Removed some unused things. Changed some cooldowns for enemy behaviour trees. Changed some attack speeds for enemies. Enemy health and speeds properly setup in data table. Blood effect properly working now for bleeding. Bloodsplosion will only happen once now on death since it could happen multiple times before.
Got rid of an unused niagara system called test.
Got rid of a blueprint that was going to be used for the spell travelling niagara but instead using a niagara system as a component on the spell projectile blueprint so not being used.

Altered some cooldowns for the behaviour trees for the enemies in various different tasks including movetos and EQSs, and melee tasks etc.

Actually changed the movement speeds and maximum / current healths of different enemy types since I forgot to do it before.
Changed the attack speeds of the melee enemies in the enemy data table so that they attack faster.

Blood effect actually works for bleeding now since I made the clamp for the radius of the blood decal to be 100 instead of 1 since the damage is really small for bleeding the radius itself was being set really low.

Now setting a timer handle for the function that destroys the owner of the health component and attached weapons so can now check if its a valid timer before calling the death function which stops multiple death functions from being called for the same actor.