New melee spell and niagara effect
Spell works and vfx work, all needs now is dmg to be specified and actually just dmg in general hitbox and the such. Although there will be a merge coming up because of conflict in weapon parent so hopefully this goes well. dont think we've had to do this yet.