A Thousand Cuts spell damage works however
it is very performance heavy cost bla bla so im gonan do some hcanges when i get back.

OKAY PROFESSIONAL TIME

The multiple traces are very perfomance heavy, as well as the blood but we fixed that by making it trigger way less.
Also the blood now has a chance to apply as a parameter on the take damage function so for all attacks you might want to change this parameter to be "1.0" so it actually applies blood.

Will make the traces happen way less and make damage work differently probably with a seperate function.