Altered camera rotation and camera max zooms as well as camera furthest player location zoom amount multiplier. Enemy boss dialogue implemented. Enemy boss trigger updated and bettered in conjunction with boss dialogue.
Increased the max zoom for the camera and also increased the amount to divide by for determining the max distance for the zoom of the camera.

Boss dialogue widget created and implemented functionality for the dialogue into the game.

Boss trigger and boss dialogue conjoining functionality implemented.
Functionality implemented for zooming in for the boss and moving the camera to be in front of the boss and then triggering the dialogue and then returning to the original position and rotation for tracking the players.