Reviving update / advancement.
Reviving is now in its final form where a player must hold down the button for a certain amount of time to revive someone and if they go outside of the range then the reviving will stop. Their movement speed is also set to be really really low so they cannot move whilst reviving and the movement of the person who is downed is also more greatly reduced now so they can barely move.

All of this resets as it should with, I think, all of the edge cases accounted for hopefully.