Optimised thousand cuts and removed bloodsplosion healing other enemies.
Thousand cuts debugging to find where it was making the game lag and then optimised to not loop through as often since it was exponentially increasing with the number of enemies hit.

Bloodsplosion now no longer loops through to detect hit enemies to heal them and will only heal the players that are within the vicinity. Also optimised this to use an array of hit actors and check if the hit actor was already within the array.