Boss enemy scaling. Fixed character select countdown. Added in the functionality for additional skeletal meshes on the player select blueprint and functionality for changing the additional skeletal meshes for the player select.
Boss 3D scale is now less for Jess' and then it lerps to be bigger for Tino's in the boss second phase transition.

Fixed the character select countdown so will only activate when all active players are ready since it worked with a player being ready after the first with both the first and second being active so that is fixed now.

Added the functionality for adding and removing the additional skeletal meshes from the player select blueprint so that they now show properly in the character select screen.

Also tried to fix the issue with the hovering for the character select ready button but couldn't get it working without messing up other stuff so may have to see about that in the future.