New character interface function for getting the main skeletal mesh of the character so not accidentally getting additional skeletal meshes. Gone through all get component by classes and replaced any that get skeletal mesh component with this function. Energy ball is now also AOE.
Created a new function in the character interface that returns a skeletal mesh component which will be used for the main skeletal mesh of the character since there is now the possibility of additional skeletal meshes so had to go through every single get component by class and see if any were getting skeletal mesh components and replace them with this function. Then setup the function in the player, enemies and boss blueprint.