Altered payload speed changes in map level. UI sounds implemented. A thousand cuts display info.
Changed pretty much all of the payload speed change amounts in the level.

Changed the sounds that play when clicking button and hovering button, and joining sound to be the sounds that Savio gave us.

Added in another field for the spell ability display info for the thousand cuts spell.