Changed stretchedness of character reps of character select screen. Fixed an issue with spamming the thousand cuts attack.
Changed the size of the character representations in the character select screen so it isn't as stretched anymore.

Also fixed an issue with the thousand cuts melee spell where spamming the attack whilst it is in process was glitching out and causing the player to not be able to move or do any actions. The fix was ... setting the animation to be looping yayayay.