Turned all traces off and hidden aoe persistence collision. Fixed issue with aoe persistence colour.  Connected all spell info missing fields when breaking and making the structs in blueprints. Bombs take longer to get to 0 speed. Sound functionality for various spell things implemented. Fixed an issue with playing the spell cooldown end animation in HUD. Added sounds to all spells (can always change just want to be able to record with good audio).
All traces are no longer visible.
AOE persistence collision is now hidden in game again.
AOE persistence colour had been set to not instance editable at some point so fixed that.

If various blueprints we have to break and remake the spell info struct to change the data within it, for example when splitting to reduce the stats of the split projectiles, so everytime a new field is added to spell info or any of its nested structs we have to update these makes/breaks so have done that now for the audio struct in the spell struct so everything should be passing correctly now.

Altered the speed to 0 for the bomb spells so that they more gradually decrease in speed until their explosion.

Added in the functionality for playing the correct sounds for various spell related blueprints: aoe, split, piercing and the spell projectile blueprint itself as well as charging.

Changing the pitch multiplier to change the duration of the sounds in blueprints: ranged attack component, and spell projectile blueprint so that they kinda match the length of the animation / spell duration. This by proxy makes the sounds sound a little different but I still think they are suitable.

Fixed an issue with the spell indicator in the HUD with it playing the spell cooldown end animation which now works for all spells that the players have yayayay.

Setup all sounds for the spells in the spells data table (can always change these around over the next day or so just want to be able to record the spells with good audio).