Kamarras Sword implemented.
Added emma's sword model, sorted out their textures and merged them and then sorted out the pivot. Also fixed some stuff with the melee weapons data table for adding this and had to change the name of the weapon in the enum. Also had to add some more checks in the animation blueprint for the attacking so that she uses the correct weapon for the attacks and doesn't switch to the katana. Added the correct sockets to the sword static mesh so that the hitbox works correctly.