Payload speed increase and health increase by 50% (2000 - 3000). Camera bounds thickened and positioned changed to avoid players dragging other players. Changed amplitude and frequency of spell camera shake and changed clamping for that in ranged attack component as well. Changed a lot of clamps surrounding camera shake to be lower maximums. Fixed bloodsplosion healing downed players. Increased player out of combat healing and payload out of combat healing and decreased out of combat delay check. Changed checking for game over for players. Fixed bug with game over. Changed spawn amounts in various enemy spawn changes in level. When payload destroyed function called game is only over if the payload is also active.
Payload speed increased from 150-225.
Payload health increased from 2000-3000.

Camera bounds have tripled in thickness and also had their positions updated to more better reflect this as well as fixing an issue with players being able to drag other players when the camera is following players (not the payload).

Amplitude and frequency of the spell shoot camera shake has been decreased and all clamps that use this camera shake have had their maximums reduced greatly to help avoid the mass amounts of camera shakes that happen when 4 people play the game.

Game over check functions differently so hopefully has fixed the issue couldn't replicate the issue in source though originally so this could be an issue with the build
:(

Game over bug fixed: if a player goes down whilst a player is in downed transition but from downed to back to up again then the game would end so now we also check if the health is <= to 0 for the game over check.

Turns out the modular balancing for amount of players was actually decent and past me was actually smart about it: the logic was that the original amount is was going to spawn was divided by (5-num of players) with a check if it is 0 and if so then its actually 1 so enemies will actually spawn. This means that if 4 enemies were supposed to spawn and 1 player was active then only 1 enemy would spawn and if 4 players were active the original amount would spawn. Whereas I thought for some reason in my head that I did it multiplicatively where 4 players active would mean 4 times the number of enemies should spawn. So to help balance I instead changed some of the amounts of enemies to spawn in the various enemy change triggers in the level.

When payload destroyed function is called the game is now only over if the payload is also active when this is called which should remove an edge case for enemies attacking and destroying the payload when in the boss fight.